Update interview_slot model
