+++ build.py
@@ -5,7 +5,7 @@ from school import *
 from city import *
 from mentor import *
 from interview_slot import *
-from interview import *
+
 
 def build_tables():
-    db.create_tables([Applicant, School, City, Mentor, InterviewSlot, Interview], safe=True)
+    db.create_tables([Applicant, School, City, Mentor, InterviewSlot], safe=True)
