+++ main.py
@@ -1,6 +1,5 @@
 from models import *
 from applicant import *
-from interview import *
 from interview_slot import *
 from school import *
 from mentor import *
